requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.4.1
    # via django
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.7
    # via requests
cryptography==35.0.0
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
dj-database-url==0.5.0
    # via environs
dj-email-url==1.0.2
    # via environs
django==3.2.8
    # via
    #   -r requirements.in
    #   django-allauth
django-allauth==0.45.0
    # via -r requirements.in
django-autocomplete-light==3.8.2
    # via -r requirements.in
django-cache-url==3.2.3
    # via environs
django-widget-tweaks==1.4.8
    # via -r requirements.in
environs[django]==9.3.4
    # via -r requirements.in
idna==3.3
    # via requests
marshmallow==3.14.0
    # via environs
oauthlib==3.1.1
    # via requests-oauthlib
pillow==8.4.0
    # via -r requirements.in
pycparser==2.20
    # via cffi
pyjwt[crypto]==2.3.0
    # via django-allauth
python-dotenv==0.19.1
    # via environs
python3-openid==3.2.0
    # via django-allauth
pytz==2021.3
    # via django
requests==2.26.0
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via django-allauth
six==1.16.0
    # via django-autocomplete-light
sqlparse==0.4.2
    # via django
urllib3==1.26.7
    # via requests
whitenoise==5.3.0
    # via -r requirements.in
